+++ test/helpers/FakeQtNavigator.js
