+++ test/helpers/FakeXMLHttpRequest.js
