+++ test/qtsyncprovider.js
@@ -5,7 +5,7 @@ var SandboxedModule = require('sandboxed-module');
 SandboxedModule.registerBuiltInSourceTransformer('istanbul');
 var QtSyncProvider = SandboxedModule.require('../lib/web3/qtsync', {
     globals: {
-        navigator: require('./FakeQtNavigator')
+        navigator: require('./helpers/FakeQtNavigator')
     }
 });
